Add feature gating for GUI and TTY backends
