+++ rust_term/Cargo.toml
@@ -5,3 +5,7 @@ edition = "2021"
 
 [dependencies]
 libc = "0.2"
+[features]
+default = ["tty"]
+tty = []
+gui = []
