+++ rust_window/Cargo.toml
@@ -9,3 +9,7 @@ once_cell = "1"
 [lib]
 name = "rust_window"
 crate-type = ["staticlib", "rlib"]
+[features]
+default = ["tty"]
+tty = []
+gui = []
